% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorful-palettes.R
\name{biasPalette}
\alias{biasPalette}
\title{Color palette for 'bias' plots.}
\usage{
biasPalette(n, zlim = c(-1, 1), col = c("dodgerblue3", "firebrick3"),
  symmetric = TRUE, p = 0.8)
}
\arguments{
\item{n}{Number of colors}

\item{zlim}{Range of the variable, zero must be inside.}

\item{col}{Vector of two colors, for the negative and positive values.}

\item{symmetric}{Boolean, if TRUE makes the palette symmetric (same intensity given the absolute value of the variable).}

\item{p}{Power to scale the increase in intensity. Default is 0.8, 1 gives linear decrease.}
}
\description{
Color palette for 'bias' plots.
}
