% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorful-palettes.R
\name{divergencePalette}
\alias{divergencePalette}
\title{Color palette for 'bias' plots.}
\usage{
divergencePalette(n = 64, zlim = c(-1, 1), col = c("dodgerblue3",
  "firebrick3"), symmetric = TRUE, p = 0.8, center = 0)
}
\arguments{
\item{n}{Number of colors}

\item{zlim}{Range of the variable, zero must be inside.}

\item{col}{Vector of two colors, for the negative and positive values.}

\item{symmetric}{Boolean, if TRUE makes the palette symmetric (same intensity given the absolute value of the variable).}

\item{p}{Power to scale the increase in intensity. Default is 0.8, 1 gives linear decrease.}

\item{center}{Center of the palette, zero by default.}
}
\description{
Color palette for 'bias' plots.
}
