% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorful-spiderplot.R
\name{spiderAxis}
\alias{spiderAxis}
\title{Title}
\usage{
spiderAxis(n, at = 0.5, rmin = 1/nvar, theta = 0, col = "grey",
  labels = NULL, clockwise = FALSE, ...)
}
\arguments{
\item{col}{}
}
\description{
Title
}
